[refactor] CardManager 클래스 삭제 #16
struct는 값 타입이라서 참조를 할 수 없음. 따라서 CardManager의 역할을 수행하지 못함 -> 새로운 인스턴스를 만들기 때문
CardManager의 역할을 Category 객체에서 하도록 변경
